requirements.txt
python-telegram-bot==20.5
scikit-learn==1.2.2
pandas==2.0.1
sqlalchemy==2.0.4
psycopg2-binary==2.9.5
nltk==3.8.1
networkx==3.1
python-crfsuite==0.9.8
torch==2.0.0       